requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
httpx==0.26.0

# Templates
jinja2==3.1.3

# Config
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
